Add more GCC warnings and fix them

+++ Framework/imgui/imgui_impl_glfw_vulkan.cpp
@@ -805,7 +805,7 @@ void ImGui_ImplGlfwVulkan_NewFrame()
 
     // Setup time step
     double current_time =  glfwGetTime();
-    io.DeltaTime = g_Time > 0.0 ? (float)(current_time - g_Time) : (float)(1.0f/60.0f);
+    io.DeltaTime = g_Time > 0.0 ? (float)(current_time - g_Time) : (1.0f/60.0f);
     g_Time = current_time;
 
     // Setup inputs
